@@ -21,14 +21,14 @@
     "immutable": "^4.3.6",
     "lodash": "^4.17.21",
     "moment": "^2.30.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-syntax-highlighter": "^16.1.0",
     "reactflow": "^11.11.4"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.9.1",
-    "@testing-library/react": "^12.1.5",
+    "@testing-library/react": "^14.3.1",
     "@testing-library/user-event": "^14.5.2",
     "@vitejs/plugin-react": "^4.2.0",
     "@vitest/ui": "^4.0.8",
